+++ src/lys/apps/user_auth/modules/user/inputs.py
@@ -24,9 +24,6 @@ class CreateUserInput:
     email: strawberry.auto = strawberry.field(
         description="Email address for the new user (will be normalized to lowercase)"
     )
-    password: strawberry.auto = strawberry.field(
-        description="Password (min 8 chars, must contain at least one letter and one digit)"
-    )
     language_code: strawberry.auto = strawberry.field(
         description="Language code in format 'en' or 'en-US'"
     )
@@ -45,9 +42,6 @@ class CreateUserInput:
 class CreateSuperUserInput:
     email: strawberry.auto = strawberry.field(
         description="Email address for the new super user (will be normalized to lowercase)"
-    )
-    password: strawberry.auto = strawberry.field(
-        description="Password (min 8 chars, must contain at least one letter and one digit)"
     )
     language_code: strawberry.auto = strawberry.field(
         description="Language code in format 'en' or 'en-US'"